Bumps LTS to 18.6 (GHC 8.10.4); Dhall 1.40.1 (standard 20.2.0); more consistent options
LTS 18.6 is the last LTS version supporting GHC 8.10.4. LTS 18.7 is
meant for GHC 8.10.6 and we need to wait for `ghc-musl` to support it.

Few options were renamed:

* `--typecheck` → `--config-typecheck`
* `--print-config-type` → `--config-print-type`

This makes all config related options more consistent allowing us to add
more options in the future.

Some typos were fixed and documentation updated. In addition to that
manual page was introduced.